--- website-2/docs/1. Introduction/1.1 About.md
@@ -1,45 +1,0 @@
----
-slug: /
----
-
-[![CircleCI](https://circleci.com/gh/JaneJeon/blink.svg?style=shield)](https://circleci.com/gh/JaneJeon/blink)
-[![Prettier code style](https://img.shields.io/badge/code_style-prettier-ff69b4.svg)](https://github.com/prettier/prettier)
-[![Deploy to Heroku](https://img.shields.io/badge/deploy%20to-heroku-6762a6)](https://heroku.com/deploy?template=https://github.com/JaneJeon/blink)
-
-> Modern, lightweight, planet-scale link shortener for teams 🎉
->
-> Easy to setup, connect with your org's SSO, hook up analytics, and extend!
-
-Blink is a link shortener that is:
-
-- _Simple and Lightweight_: the simple "base" architecture means you can literally host it for free, and maintenance is easy.
-- _Planet-Scale_: you can host Blink alone for simplicity, or put a CDN in front of it to scale out Blink and easily plug in your own analytics!
-- _For Teams_: external identity/SSO via OIDC/OAuth support is baked right into Blink; no need to pay $$$$$ or hack together SSO support via reverse proxies! Plus, baked in access control (RBAC)!
-- _Extensible_: not only does it have a REST API, but it also allows you to integrate it with public applications by protecting API calls with OAuth2!
-
-## Screenshots
-
-<figure>
-  <img
-  src="/img/short.png"
-  alt="Link shortening"/>
-  <figcaption>Link shortening</figcaption>
-</figure>
-<figure>
-  <img
-  src="/img/links.png"
-  alt="Link management"/>
-  <figcaption>Link management</figcaption>
-</figure>
-<figure>
-  <img
-  src="/img/users.png"
-  alt="User management"/>
-  <figcaption>User management</figcaption>
-</figure>
-
----
-
-_For development, see the GitHub repo's [README](https://github.com/JaneJeon/blink)_
-
-This project is based off of [Express-Objection-Template](https://github.com/JaneJeon/express-objection-template), and heavily uses [objection-authorize](https://github.com/JaneJeon/objection-authorize) and [objection-hashid](https://github.com/JaneJeon/objection-hashid) to drive much of "hidden" logic and make isomorphism possible.
