--- website-2/docs/1. Introduction/1.4 FAQs.md
@@ -1,32 +1,0 @@
-## Do I need an external identity provider?
-
-See the [Prerequisites page](/Installation/2.1%20Prerequisites).
-
-## Why a URL Shortener?
-
-Besides the obvious (shorter URLs and the ability to "brand" links), link shorteners can serve as “middleware” to essentially “decouple” the source and the destination. And with that decoupling comes a whole host of benefits:
-
-- the ability to gain analytics over links clicked, who is clicking it, etc (pluggable)
-- the ability to enforce HTTPS (which Blink already does)
-- the ability to healthcheck links in one convenient place (planned)
-- the ability to mass-rename links (such as if you moved your domain) without breaking existing, shortened links (planned)
-- the ability to rename a longer link into a form that’s more readable
-- etc, etc.
-
-## But URL shorteners are "bad" for the web!!1!
-
-Websites “rot” regardless. Millions of links die every day. Domains change. Entire sites are shut down, or their URL format completely changes. Stuff gets paywalled, or simply gets deleted.
-
-To say you shouldn’t try to actively manage (i.e. keep alive) links - all from different websites you can’t control - via link shorteners because “they die anyway once the service goes down” is a bit defeatist, no?
-
-## Does this image include a proxy?
-
-No. This application only speaks HTTP. To setup HTTPS for your self-hosted installation, please see [here](/Installation/2.3%20Advanced)
-
-## Does it give you analytics like Bitly?
-
-Not off the bat - the "raison d'etre" of Blink is to decouple analytics from the link shortener. However, setting up analytics for Blink can be [extremely simple](/Installation/2.2%20Shortcut).
-
-## Do you have to use a CDN? What about muh privacy?
-
-You do not _have_ to use a CDN. See the [Architecture page](/Introduction/1.3%20Architecture) for how else it can be used (hint: it can be used "bare" or with any caching layer of your choosing).
