--- website-2/docusaurus.config.js
@@ -1,130 +1,0 @@
-// @ts-check
-// Note: type annotations allow type checking and IDEs autocompletion
-
-const lightCodeTheme = require('prism-react-renderer/themes/github')
-const darkCodeTheme = require('prism-react-renderer/themes/dracula')
-
-/** @type {import('@docusaurus/types').Config} */
-const config = {
-  title: 'Blink',
-  tagline: 'Modern, lightweight, planet-scale link shortener for teams',
-  favicon: 'img/favicon.ico',
-
-  // Set the production url of your site here
-  url: 'https://docs.blink.rest',
-  // Set the /<baseUrl>/ pathname under which your site is served
-  // For GitHub pages deployment, it is often '/<projectName>/'
-  baseUrl: '/',
-
-  // GitHub pages deployment config.
-  // If you aren't using GitHub pages, you don't need these.
-  // organizationName: 'facebook', // Usually your GitHub org/user name.
-  // projectName: 'docusaurus', // Usually your repo name.
-
-  onBrokenLinks: 'throw',
-  onBrokenMarkdownLinks: 'warn',
-
-  // Even if you don't use internalization, you can use this field to set useful
-  // metadata like html lang. For example, if your site is Chinese, you may want
-  // to replace "en" with "zh-Hans".
-  i18n: {
-    defaultLocale: 'en',
-    locales: ['en']
-  },
-
-  presets: [
-    [
-      'classic',
-      /** @type {import('@docusaurus/preset-classic').Options} */
-      ({
-        docs: {
-          sidebarPath: require.resolve('./sidebars.js'),
-          // Please change this to your repo.
-          // Remove this to remove the "edit this page" links.
-          editUrl: 'https://github.com/JaneJeon/blink/edit/master/website/',
-          routeBasePath: '/'
-        },
-        theme: {
-          customCss: require.resolve('./src/css/custom.css')
-        }
-      })
-    ]
-  ],
-
-  themeConfig:
-    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
-    ({
-      colorMode: {
-        respectPrefersColorScheme: true
-      },
-      // Replace with your project's social card
-      // image: 'img/docusaurus-social-card.jpg',
-      navbar: {
-        title: 'Blink',
-        // logo: {
-        //   alt: 'My Site Logo',
-        //   src: 'img/logo.svg'
-        // },
-        items: [
-          {
-            to: '/',
-            activeBasePath: 'docs',
-            label: 'Docs',
-            position: 'left'
-          },
-          {
-            href: 'https://github.com/JaneJeon/blink',
-            position: 'right',
-            className: 'header-github-link'
-          }
-        ]
-      },
-      footer: {
-        style: 'dark',
-        links: [
-          // {
-          //   title: 'Docs',
-          //   items: [
-          //     {
-          //       label: 'Tutorial',
-          //       to: '/docs/intro'
-          //     }
-          //   ]
-          // },
-          // {
-          //   title: 'Community',
-          //   items: [
-          //     {
-          //       label: 'Stack Overflow',
-          //       href: 'https://stackoverflow.com/questions/tagged/docusaurus'
-          //     },
-          //     {
-          //       label: 'Discord',
-          //       href: 'https://discordapp.com/invite/docusaurus'
-          //     },
-          //     {
-          //       label: 'Twitter',
-          //       href: 'https://twitter.com/docusaurus'
-          //     }
-          //   ]
-          // },
-          // {
-          //   title: 'More',
-          //   items: [
-          //     {
-          //       label: 'GitHub',
-          //       href: 'https://github.com/facebook/docusaurus'
-          //     }
-          //   ]
-          // }
-        ],
-        copyright: `Copyright © ${new Date().getFullYear()} Jane Jeon`
-      },
-      prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme
-      }
-    })
-}
-
-module.exports = config
