--- website/.npmrc
@@ -1,2 +1,0 @@
-audit=false
-fund=false
