+++ website/src/components/HomepageFeatures/index.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
-import clsx from 'clsx';
-import styles from './styles.module.css';
+import React from 'react'
+import clsx from 'clsx'
+import styles from './styles.module.css'
 
 type FeatureItem = {
-  title: string;
-  Svg: React.ComponentType<React.ComponentProps<'svg'>>;
-  description: JSX.Element;
-};
+  title: string
+  Svg: React.ComponentType<React.ComponentProps<'svg'>>
+  description: JSX.Element
+}
 
 const FeatureList: FeatureItem[] = [
   {
@@ -17,7 +17,7 @@ const FeatureList: FeatureItem[] = [
         Docusaurus was designed from the ground up to be easily installed and
         used to get your website up and running quickly.
       </>
-    ),
+    )
   },
   {
     title: 'Focus on What Matters',
@@ -27,7 +27,7 @@ const FeatureList: FeatureItem[] = [
         Docusaurus lets you focus on your docs, and we&apos;ll do the chores. Go
         ahead and move your docs into the <code>docs</code> directory.
       </>
-    ),
+    )
   },
   {
     title: 'Powered by React',
@@ -37,11 +37,11 @@ const FeatureList: FeatureItem[] = [
         Extend or customize your website layout by reusing React. Docusaurus can
         be extended while reusing the same header and footer.
       </>
-    ),
-  },
-];
+    )
+  }
+]
 
-function Feature({title, Svg, description}: FeatureItem) {
+function Feature({ title, Svg, description }: FeatureItem) {
   return (
     <div className={clsx('col col--4')}>
       <div className="text--center">
@@ -52,7 +52,7 @@ function Feature({title, Svg, description}: FeatureItem) {
         <p>{description}</p>
       </div>
     </div>
-  );
+  )
 }
 
 export default function HomepageFeatures(): JSX.Element {
@@ -66,5 +66,5 @@ export default function HomepageFeatures(): JSX.Element {
         </div>
       </div>
     </section>
-  );
+  )
 }
